sentiment-analyzer/requirements.txt
beautifulsoup4==4.12.3
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
nltk==3.8.1
psycopg2==2.9.9
python-dotenv==1.0.1
regex==2023.12.25
soupsieve==2.5
SQLAlchemy==2.0.28
textblob==0.18.0.post0
tqdm==4.66.2
typing_extensions==4.10.0
Werkzeug==3.0.1
